Moved debug_throw to rj_utils/logging.hpp
