+++ rj_common/include/rj_common/utils.hpp
@@ -5,38 +5,9 @@
 
 #include <spdlog/spdlog.h>
 
-const static bool kThrowDebugExceptions = true;
-
 template <typename T>
 inline int signum(T val) {
     return (0 < val) - (val <= 0);
-}
-
-inline void debug_log(const std::string& e) { spdlog::debug(e); }
-
-inline void debug_log(const std::exception& e) { spdlog::debug(e.what()); }
-
-inline void debug_log_if(const std::string& e, bool condition) {
-    if (condition) {
-        debug_log(e);
-    }
-}
-
-template <class T,
-          typename std::enable_if<std::is_base_of<std::exception, T>::value, int>::type = 0>
-inline void debug_throw(const T& e) {
-    debug_log(e);
-    if (kThrowDebugExceptions) {
-        throw e;
-    }
-}
-
-inline void debug_throw(const std::string& string) { debug_throw(std::runtime_error(string)); }
-
-inline void debug_throw_if(const std::string& string, bool condition) {
-    if (condition) {
-        debug_throw(std::runtime_error(string));
-    }
 }
 
 /**
